Split fixup.go into pieces
